move server logic related to enacting policy to new file
